src/api/requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
PyMuPDF
nest_asyncio
python-dotenv
python-multipart